Revise times of checking mgmt port link status to 100
The polling loop of checking link status is a total 5 seconds in the
worst case. It may be timeout eventually if the link partner has
capibility of 1000FD.

I added a debug message to show total times of checking the link status
of mgmt port.

    [ -r $operstate ] && while [ $i -lt 100 ] ; do
        if [ "$(cat $operstate)" = "up" ] ; then
            _log_info_msg "up. (waited $i times)\n"
            return 0
        fi
        sleep 0.1
        i=$(( $i + 1 ))
    done

If both mgmt port and its link partner have 1000FD capability with
auto-negotiation enabled, they will spend more time to accomplish the
negotiation.

We did hundreds of testing on Accton AS5710_54x. The test result are:

- 40 times (55%)
- 50 times (45%)
- 69 times (8%)
- 79 times (2%)

It indicates most of time mgmt port is link-up in 5 seconds (40 times
and 50 times). It has 10% of probability over 5 seconds (69 times and
79 times). So that we decided to enlarge the total seconds to 10 i.e.,
100 times.