Support for embedding arbitrary string to Linux kernel command line
Issue #45:

This commit adds supports for adding arbitrary strings to the Linux
kernel commandline.

A user can set a shell env variable USER_GRUB_CMDLINE_LINUX with a
string that needs to be embedded in the kernel commandline.

Eg:
export USER_GRUB_CMDLINE_LINUX="install_url=http://myserver/path/to/onie-installer"

and the string will get embedded into the GRUB config so that it's
passed to it when loading the ONIE Linux kernel.