busybox udhcpc: Make hostname sanitization be optional
The hostname leading with digit passed by DHCP opt 66 is invalid
because it is constrained by RFC-1035.

As mentioned in networking/udhcp/Config.src:
    udhcpc will check some options (such as option 12 -
    hostname) and if they don't look like valid hostnames
    (for example, if they start with dash or contain spaces),
    they will be replaced with string "bad" when exporting
    to the environment.

The invalid hostname will be replaced with string "bad".  It causes
ONIE use the "bad" hostname to fetch installer.  This patch will
disable the check and pass TFTP server name to ONIE without any
alteration.

Ref:
http://git.busybox.net/busybox/commit/?id=85090c162b322a4ffe53d251e59bbfc212a829ee

Signed-off-by: david_yang <david_yang@accton.com>