Add copyright notices to all files
To help the Open Compute Project protect the openness of the ONIE
project, proper copyright notices need to be in place for all the
applicable files.

This patch adds a copyright notice to most files based upon the
authors of the files as reported by 'git log'.  The format of the
notice is:

  #  Copyright (C) 2014 ONIE Developer <onie_helper@onie.org>
  #
  #  SPDX-License-Identifier:     GPL-2.0

The SPDX-License-Identifier is a short hand way to reference the
complete text of the GPLv2 license, which ONIE has used from the
beginning.  See the COPYING file for more on that.

  http://spdx.org/licenses/GPL-2.0