xtools: restart build on failure
Today the crosstools-ng build has the bad property that when
restarting a build after a failure the crosstool-ng system believes
the toolchain is built correctly.  This leads to all manner of
downstream build errors and confusion for developers who are trying to
make progress after a crosstool-ng failure.

This patch coaxes the xtools target into being "out of date" whenever
the crosstool-ng build fails.  This is accomplished by removing the
$(XTOOLS_BUILD_DIR)/.config.2 file, which the crosstool-ng system uses
as a prerequisite for building, whenever the crosstool-ng build fails.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>