x86: include the recovery ISO into the 'all' Makefile target
Reviewed By: Carlos Cardenas

When building the x86 based machine types it is often useful to have
the recovery ISO, though not strictly required all the time.

The build time of the recovery ISO is small compared to everything
else, so including it by default in the 'all' Makefile target has
little impact.

This patch adds building the recovery ISO to the 'all' Makefile target
for x86 machine.

Testing Done:

Compiled the kvm_x86_64 machine using the 'all' target.  Verified the
build was successful, the recovery ISO was created and nothing broke.

Compiled the dell/dell_s6000_s1220 machine using the 'all' target.
This machine is noteworthy because it's machine.make *already*
includes the building of the pxe-efi64 recovery image, which has the
recovery ISO as a dependency.  Verified the build was successful, the
recovery ISO was created and nothing broke.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>