x86_64-uefi: add support for efibootmgr tool
When UEFI support is enabled on x86_64, this patch adds support for
the efibootmgr CLI utility.  This utility pulls in dependencies on
lib-efivar and lib-pci, so those projects are also added to the
system.

For efibootmgr the base is taken from the public 0.11.0 release.
Added to that are the 8 efibootmgr github commits made since the
0.11.0 release.

For efivar the base is taken from the public 0.15 release.  Added to
that are the xx efivar github commits made since the 0.15 release.

Testing Done:

Created an ONIE rescue ISO and booted it on an UEFI capable machine.
Upon boot up ran the 'efibootmgr' CLI and verified the output.