Accton platforms: be able to install on machines running ONIE v1.6
The platforms are including:
- AS4600_54T
- AS5600_52X
- AS5610_52X
- AS5700_96X
- AS5710_54X
- AS5712_54X
- AS6700_32X
- AS6701_32X
- AS6710_32X
- AS6712_32X
- AS7700_32X
- AS7702_32X

The machine name in ONIE v1.6 did not lead with '<VENDOR>_' prefix
until v2014.08.  Hence, The legacy ONIE can't upgrade to recent
version easily.  For instance, updater for 'accton_as5610_52x' can't
install on 'as5610_52x' machine because of the names not the same.

This patch makes recent updater be able to install on the machines
running legacy ONIE.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>