x86: ISO image creation improvements
This patch improves and expands the capabilities of the ONIE recovery
ISO image.

The ISO image is now usable in the following contexts:

- legacy BIOS CD-ROM
- legacy BIOS hard disk (USB drive)
- UEFI CD-ROM
- UEFI hard disk (USB drive)

In an attempt to clean things up, the ISO image no longer depends on
the SYSLINUX project.  All of the boot scenarios are now handled by
GRUB2.  This simplifies the project build dependencies.

The ISO image was exercised using the x86_64-kvm_x86_64 virtual
machine and QEMU as follows:

  $ qemu -cdrom  onie-recovery-x86_64-kvm_x86_64-r0.iso                 <-- legacy BIOS
  $ qemu -hda    onie-recovery-x86_64-kvm_x86_64-r0.iso                 <-- legacy BIOS
  $ qemu -hda    onie-recovery-x86_64-kvm_x86_64-r0.iso -bios OVMF.fd   <-- UEFI
  $ qemu -cdrom  onie-recovery-x86_64-kvm_x86_64-r0.iso -bios OVMF.fd   <-- UEFI

All four scenarios boot correctly.