fsl_p2020rdbpca: general updates and use kernel 4.1.1
The demonstration PowerPC platform, fsl_p2020rdbpca, has gone a long
time without any updates.  This patch moves the fsl_p2020rdbpca
machine forward in a few areas:

- use kernel 4.1.1
- use uClibc 0.9.33.2
- add the busybox onie-syseeprom command (implemented in NOR flash)

One interesting thing to note is that the size of the 4.1.1 kernel
image is considerablely larger than the 3.2.X image, even when
configured more or less the same.  The .itb files comes in at about
4.5MiB, where as with 3.2.X is was about 3.5MiB.

The size increase causes the NOR partition layout to change a little
bit, essentially shrinking the size of the "open" partition and
growing the size of the "onie" partition from 4MB to 6MB.

Another interesting thing with kernel 4.1.1 and device tree is that
/proc/device-tree is now a symlink to /sys/firmware/devicetree/base.
This required a small change to
rootconf/powerpc-softfloat/sysroot-lib-onie/init-arch to accommodate
the difference.

Testing Done:

Built the demonstration PowerPC machine fsl_p2020rdbpca and verified:

- boots OK
- loads the demo OS OK
- loads the demo diag OK
- ONIE update works OK

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>