x86: add flashrom support
This patch adds the 'flashrom' utility and pciutils library to the
ONIE initramfs for x86 builds.  This tool is *very* handy on x86
systems for updating the BIOS firmware.

Adding the utility and library increases the size of the kvm_x86_64
image by ~100,000 bytes.

By default most flashrom methods except for 'internal' and Intel NIC
EEPROMs are disabled from the build.  Individual platforms can define
FLASHROM_MAKE_CONFIG_PLATFORM to enable support for specific flashing
methods, for example:

  FLASHROM_MAKE_CONFIG_PLATFORM = CONFIG_XYZ=yes

See the Makefile in the flashrom source directory for the complete
list of CONFIG_XYZ variables.

Note: Leaving the Intel NIC EEPROMs enabled, as that could be useful
in manufacturing to program the MAC address for eth0.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>