Add tests for using well-known server name
Add tests for using a well-known server name ('onie-server')
for #219.

These 20 tests ensures that an ONIE client can perform
image discovery using a well known server name ('onie-server')
with HTTP and TFTP.

This commit also adds a new service, 'dns', and allows dnsmasq
to be configured as either a standalone DNS server to be used
with ISC-DHCP-Server or with other dnsmasq services.

Tested on KVM image.