grub: fix race condition in grub build
When running 'make all' with -j24 occasionally the build would fail
for the grub-install and grub-host-install targets.

The grub builds share a common source directory and are supposed to
use separate output directory.  That part was working fine.

The trouble was the build of the grub info pages (documentation)
happens in the source directory, instead of the separate binary output
directories.  The parallel builds would clobber each other building
the documentation.

The fix here is to remove the building of the grub documentation,
which we are not using anyway.

Testing:

Before the fix building an x86 machine from scratch would fail 1 out
of 5 times or so.  After the fix *all* x86 machines built from scratch
successfully.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>