kvm_x86_64: allow some flexibility for install disk
Previously the kvm_x86_64 machine defaulted to using /dev/vda for
installing ONIE.  This was fine as long as the VM is configured to use
a virtual IO block device for the storage medium.  This works great
with QEMU.

Some hypervisors do not support virtual IO block devices.  However,
SATA controller are commonly supported.  In order to be more flexible
in the choice of hypervisor this patch let's the installer choose
either /dev/vda or /dev/sda at runtime for installing ONIE.

Testing:

Installed the ONIE image using a virtio disk controller and using a
SATA disk controller.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>