armv7a: add build-system support for ARM
This patch adds support for an armv7a soft-float (no FPU) GCC
toolchain using uClibc.

The ARM platform is following the PowerPC model of using U-Boot with a
kernel, initramfs and dtb file in a .itb U-Boot image file.

In several areas the build system is refactor to treat PowerPC and ARM
in a similar manner.