Accton x86_64 platform: add management port work around to kernel driver
The platforms are including:

- AS5512_54X
- AS7512_32X
- AS7712_32X

When doing power cycle on the machine repeatly, it may cause management
port out of control.  The probability is around 0.4%~0.7%.  However, it
will recover in the next power cycle.  The issue also occurred in recent
Intel igb driver.  This work around intends to avoid the issue to occur.

NOS running on the machines should also have the work around.