grub: support u-boot-like ONIE boot commands
In Accton, the QA/MFG engineers think that current GRUB menu is hard for
auto test program to select specific boot target.  The GRUB menu may be
starting at either ONIE or DIAG page.  The program can not easily predict
which page will be the first one.  And it does not easily move between the
pages.

The patch supplies ONIE boot commands that can be performed in the ONIE
GRUB shell directly to boot target.  It is useful for test without playing
the mouse maze algorithm.

The patch is compatible with present ONIE design and has been tested in
Accton AS7712_32X.