install/update messages cleanup.
Fixing #316

On Interface Masters platforms /boot is shared for
ONIE and NOS grub files. Hence scripts are doing back up of
content if any. On busybox wild card characters are specific
and complaining on '*' in empty directory.
Fix is to move output of 'mv' command to /dev/null.
Functionality is still correct for backing up and restoring grub
content.

Signed-off-by: Vitaliy Ivanov <vitaliyi@interfacemasters.com>
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>