grub: fix build time race condition
A build time race condition exists for the grub package in the
following circumstances:

- Using parallel make (make -j4 for example)
- Enabling UEFI (or Coreboot) support

When enabling UEFI/Coreboot the grub package is built twice and
installed into the same development sysroot.  A race condition exists
during the install step for files common to i386 and UEFI/coreboot.

The fix here is to have the grub-uefi and grub-coreboot install
Makefile targets depend on the grub-i386 install target, thus
serializing the install and removing the race condition.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>