firmware updates: add infra for building firmware updates
The primary goal here is to provide a relatively easy mechanism for
machine maintainers to generate a "firmware update" for their
machines.

A firmware update shares many properties with an ONIE update image as
implemented previously in commits
a8e712b451dce0f40c8a9ce117922be758ba9a38 and
477cd471415102466ee86226263eb4a5c23a9ec3:

  commit a8e712b451dce0f40c8a9ce117922be758ba9a38
  Author: Curt Brune <curt@cumulusnetworks.com>
  Date:   Tue Jun 30 10:56:47 2015 -0700

      pending firmware update discovery mechanism

  commit 477cd471415102466ee86226263eb4a5c23a9ec3
  Author: Curt Brune <curt@cumulusnetworks.com>
  Date:   Mon Jul 13 14:39:38 2015 -0700

      x86 firmware update: add onie-fwpkg CLI tool

The secondary goal is to treat ONIE update images and firmware update
images in a similar fashion.

This patch introduces a new Makefile variable, FIRMWARE_UPDATE_ENABLE,
that machine maintainers set to "yes" in their machine.make file.
When set, this enables a new Makefile target "firmware-update" that
generates a firmware update image based on machine specific files.

When firmware updates are enabled, the following files must exist in
the machine's "firmware" directory, a new sub-directory of the
platform specific directory (the same directory as the machine.make
file):

  platform/machine.make
  platform/firmware/fw-version.make
  platform/firmware/fw-install.sh

fw-version.make -- this is small Makefile fragment that defines the
firmware update version by setting the FW_VERSION variable.  This
version is a free form string up to the machine maintainer.

fw-install.sh -- this is the firmware update entry point shell script,
provided by the machine maintainer.  The firmware update
infrastructure calls this script to start the firmware update process.

What fw-install.sh does is completely up to the machine maintainer.

The platform/firmware/ directory can contain additional files and
directories, depending on the needs of the machine maintainer.

When fw-install.sh executes the current working directory is the same
directory as the script.  For example consider a machine firmware
directory that contains a program called "update-bits" in a
sub-directory called "more-stuff".  The firmware directory in this
example looks like this:

  platform/firmware/fw-version.make
  platform/firmware/fw-install.sh
  platform/firmware/more-stuff/update-bits

The fw-install.sh script can refer to this program as:

  ./more-stuff/update-bits

To build the firmware update for a machine, using the
'firmware-update' Makefile target.  When complete, the firmware update
image is available in:

  build/images/onie-firmware-$(PLATFORM).bin

The output file name can be overridden by setting the
FIRMWARE_UPDATE_BASE variable on the 'make' command line.

This image can now be used with the onie-fwpkg command.

This patch also unifies the creation of ONIE updater images with
firmware update images.  Both types of images share a common generator
script, onie/build-config/scripts/onie-mk-installer.sh and other
common files from onie/installer.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>