dni_3048up: SRAM u-boot recovery image
This patch adds support for creating a u-boot recovery image that can
run out of the SoC's internal SRAM without the DDR.  This is useful in
situations where the u-boot image needs to be loaded by a JTAG
debugger.

This functionality is entirely cloned from the Accton AS4610-54
machine, from commit 17fbb7adc50e514f7f5b8373dfded362fa34a026:

  commit 17fbb7adc50e514f7f5b8373dfded362fa34a026
  Author: Vidya Sagar Ravipati <vidya@cumulusnetworks.com>
  Date:   Mon Oct 26 14:51:45 2015 -0700

      arm: accton AS4610-54: ONIE support for accton as4610-54 board

A new option is added to include/configs/DNI_3048UP.h called
CONFIG_DNI_3048UP_RECOVERY, which is disabled by default.

When this option is enabled the machine is configured to build a SRAM
based U-Boot image.

Here is an example of using this image for recovery.  The JTAG
debugger in this example is an Abatron BDI3000.  First connect to the
machine with the debugger and load the recovery u-boot image:

  BDI3000> reset halt
  BDI3000> load 0x1b008000 bdi/dni-3048UP/recovery-u-boot.bin BIN
  BDI3000> go 0x1b008000

On the serial console you should see the box go through it's paces,
possibly re-calibrating the shoom.  At the end you should have a
u-boot prompt.

Note: The recovery u-boot image is really stripped down so that it
will fit into the internal SRAM of the iProc SoC.  Not a whole lot of
functionality here.

Now load a normal u-boot image (without CONFIG_DNI_3048UP_RECOVERY
defined) into the RAM using the BDI.  This image should be larger than
the recovery image:

  BDI3000> halt
  BDI3000> load 0x61000000 bdi/dni-3048UP/u-boot.bin BIN
  Loading bdi/dni-3048UP/u-boot.bin , please wait ....
  - 830760 bytes loaded
  Loading program file passed
  BDI3000> go

Note the file size in bytes (830760).  Convert that to hex, in this
case 0xcad28.  Round up to even boundary, in this case 0xd0000.

Now use the u-boot CLI to copy that image into the SPI ROM.

  u-boot> sf probe 0
  SF: Detected MX25L25635E with page size 4 KiB, total 32 MiB
  u-boot> sf erase 0 0xd0000
  u-boot> sf write 0x61000000 0 0xd0000

Now reset and reboot:

  u-boot> reset
  resetting ...

  U-Boot SPL 2012.10-g9717418 (Jun 24 2016 - 17:33:14)
  DCO:   89 (PASS)
  HWRev: 0x61 AVS: 0x0 VOUT Init: 0x10 VOUT Set: 0x10
  ...

Now the system is functional again.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>