dni_3048up: Fix up usable SDRAM starting address
The iProc U-Boot maps physical SDRAM to begin at 0x6000_0000.

The initial commit for this platform,
65bcc5f5d2027f7fcc929959770a028c46d6352a, defines the first 16MB
(0x0100_0000) as "reserved for custom use" and excludes it from the
available free memory.

  commit 65bcc5f5d2027f7fcc929959770a028c46d6352a
  Author: vincent0083 <Vincent.Liu@agemasystems.com>
  Date:   Thu May 26 11:02:59 2016 +0800

      Add support for DNI 3048UP

None of the PowerPC ONIE U-Boot implementations do this.

This is just a waste of 16MB.

This patch sets the reserved SDRAM size to 0x0.

This also has the nice property of aligning the start of RAM to a
512MB boundary, which simplifies the memory layout for loading kernels.

For consistency, this patch also adjusts the /memory node in the
kernel .dts file to reflect the new starting memory address.

Closes:
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>

tweak: dni_3048up: Fix up usable SDRAM starting address -- Device tree