busybox: Add CONFIG_SYS_EEPROM_I2C_MEM_ADDR_BITS to machine configs
After moving forward to a new version of busybox and base
configuration file, we now require that machines explicitly state the
number of bits for i2c EEPROM addressing.  Previously the default was
8 bit and now that must be explicitly stated in the machine specific
busybox configuration file.

This patch adds the following configuration line to a number of
machines that were previously relying on the default:

  CONFIG_SYS_EEPROM_I2C_MEM_ADDR_BITS=8

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>