images.make: include certain libraries only if necessary
Commit 44a1f46a758093338cc6450796bd1a72b27484b7 introduced a
regression where some libraries were unnecessarily included in the
final sysroot:

  commit 44a1f46a758093338cc6450796bd1a72b27484b7
  Author: Pankaj Bansal <pankaj.bansal@nxp.com>
  Date:   Sun Nov 6 20:40:04 2016 +0530
   
      Update Crosstool-NG from 1.21.0 to support Aarch64

This causes problems for some platforms that have very limited storage
space.

This patch reintroduces checks to see if the libraries are needed,
instead of just including them by default.

This is only done for the uClibc case, as it is assumed glibc users
are unconcerned about a few extra 100K bytes.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>