accton_as56xx_52x: remove strace utility
The recent move to the latest version of crosstool-ng has increased
the overall image size slightly.  These two platforms have a
constrained image size and now the ONIE image no longer fits:

- accton_as5600_52x
- accton_as5610_52x

This patch removes the development debug utility, 'strace', from the
sysroot reducing the image size.  If this utility is needed for
development it can be copied up via the network or transferred using
USB.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>