busybox: re-enable CONFIG_BUSYBOX
When moving the busybox configuration from busybox-1.20.2 to
busybox-1.25.1 the CONFIG_BUSYBOX feature was accidentally dropped.

This feature was never explicitly stated in the configuration before.
Assuming it was being enabled as a dependency of some other feature
previously.

Without this feature the "busybox" command with no arguments does not
work and exits with a failure code.  Some scripts use this command to
find the version of busybox.  The command is also useful interactively
to determine what applets are installed.

This patch adds CONFIG_BUSYBOX=y to the global busybox configuration.

Closes #454
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>