installer: share files with ONIE run-time
In order to reduce code replication it is desirable to share code
between the ONIE *installer* and the ONIE *run-time*.

For example on x86_64 this is already the case for
rootconf/grub-arch/sysroot-lib-onie/onie-blkdev-common, where we want
to have all the block device information available to both the
*installer* and the ONIE *run-time*.

This patch adds a general mechanism, so that platforms can share code
in a similar manor.

For example, to share a shell script fragment between the run-time and
the installer do this:

Step 1.  Create the run-time shell script fragment.  Call it
"platform-stuff".

  $(MACHINEDIR)/rootconf/sysroot-lib-onie/platform-stuff

This fragment can be sourced at run time (say from a boot init script)
like this:

  . /lib/onie/platform-stuff

Step 2. Include the fragment in the installer.

Modify $(MACHINEDIR)/machine.make to include:

  UPDATER_IMAGE_PARTS_PLATFORM = $(MACHINEDIR)/rootconf/sysroot-lib-onie/platform-stuff

This fragment can be sourced at install time from
$(MACHINEDIR)/installer/install-platform like this:

  . ./platform-stuff

Generally, the files specified in $(UPDATER_IMAGE_PARTS_PLATFORM) are
copied into the installer image and are available in the current
working directory.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>