kvm_x86_64: test the installer/runtime code sharing
This patch adds a file,
$(MACHINEDIR)/rootconf/sysroot-lib-onie/test-install-sharing, to
exercise the installer/runtime code sharing mechanism.

If the code sharing mechanism is working properly, then the installer
will print "Success: test_install_sharing mechanism".

This code is not intended to be included in machines by default.  It
is only an example of the code sharing mechanism.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>