networking: bring up link on the loopback interface
The boot time networking script never brought the link up for the
loopback interface (lo).  This is a historical oversight that never
caused any problems with the ONIE services.

While not fatal, it is confusing that the loopback address, 127.0.0.1,
and the host name "localhost" do not function properly.

This patch brings up the "lo" interface as part of the system
networking initialization.

Closes: #498
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>