Set the UEFI boot loader statically
The 50_onie_grub GRUB fragment is intended to be executed by a NOS as
part of the NOS's /etc/grub.d scripts.

For UEFI systems this fragment has a dependence on the the ONIE file
/etc/machine.conf for the definition of the UEFI boot loader.
However, when the NOS is running ONIE's /etc/machine.conf is
unavailable.

For ONIE today, the UEFI boot loader has two possible values:

- grubx64.efi
- grubaa64.efi

These names are defined in the CPU architecture dependent Makefile
fragments in build-config/arch/<arch>.mk in the $(GRUB_IMAGE_NAME)
variable.

This patch passes the $(GRUB_IMAGE_NAME) value all the way through the
build system and sets the value statically in 50_onie_grub.

Closes: #505
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>