updater: store U-Boot version string reliably
Part of the ONIE update process is to set a few U-Boot environment
variables identifying the release of ONIE. This happens in the
update_uboot_env() function, which is defined in
installer/u-boot-arch/install-arch.

The current method for locating the version information within the
U-Boot binary only worked reliable for PowerPC systems.

This patch introduces a method that works on both PowerPC and 32-bit
ARM.

Close: #519
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>