updater: rename U-Boot env variable 'ver' to 'onie_uboot_version'
Part of the ONIE update process is to set a few U-Boot environment
variables identifying the release of ONIE. This happens in the
update_uboot_env() function, which is defined in
installer/u-boot-arch/install-arch.

One such variable is 'ver', which is intended to contain the U-Boot
build time version string.

Unfortunately the 'ver' variable is a bit special, in that at U-Boot
runtime this variable is set dynamically by U-Boot.  This can lead to
confusion when printing the value using "printenv ver" from the U-Boot
CLI compared to using "fw_printenv ver" from the ONIE CLI.

This patch renames the U-Boot variable 'ver' to 'onie_uboot_version',
storing the value in the ONIE variable namespace.  For backward
compatibly with NOSes that may be looking for 'ver', continue to set
'ver' also.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>