U-Boot: store version and build date in default environment
This patch stores the U-Boot version string and the U-Boot build date
in the default U-Boot environment block.

The default environment block is used as the initial environment
variables on a blank system.  The default environment block is also
used to recover when the environment data becomes corrupted.

The variables stored are:

- onie_build_date
- onie_uboot_version
- ver

Note 'ver' and 'onie_uboot_version' are identical.
'onie_uboot_version' is the preferred variable name, while 'ver' is
maintained for backward compatibly.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>