networking: reduce the verbosity of udhcpc
The current udhcpc client from busybox logs informational messages to
stderr as well as syslog.  The stderr logging is annoying as it
clutters the console.

This patch redirects the udhcpc stderr to /dev/null.  The message are
still logged via syslog to /var/log/messages, so no information is
lost.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>